Use bivariate exponential decay to model thresholds
Within the bounds of the SNP density and the sample size studied the
threshold function has a bivariate exponential decay distribution
shape. It would be better to simply use this function directly.
This means only two parameters need to be estimated (lambda1 and
lambda2, the decay parameters for SNP and sample size).

It also means it could be approximately linear at low densities and
asymptoting at high densities.

We shall see how this function behaves under non-linear least squares
estimation.